Split home models into multiple files
 - in order for this to work I nedeed to update some old migrations
   with the new snippets package.
