+++ home/models/__init__.py
@@ -1,0 +1,5 @@
+from . import snippets  # noqa:F401
+from . import blocks  # noqa:F401
+from . import base_pages  # noqa:F401
+from . import manual_pages  # noqa:F401
+from . import settings  # noqa:F401
